+++ atomize/device_modules/Tektronix_4000_Series.py
@@ -415,7 +415,7 @@ class Tektronix_4000_Series:
                         if ch in self.channel_dict:
                             flag = self.channel_dict[ch]
                             if flag[0] == 'C' and int(flag[-1]) <= self.analog_channels:
-                                self.device_write(str(flag) + ':SCAle ' + str(int(val/coef)))
+                                self.device_write(str(flag) + ':SCAle ' + str(float(val/coef)))
 
             elif len(channel) == 1:
                 ch = str(channel[0])
@@ -471,7 +471,7 @@ class Tektronix_4000_Series:
                         flag = self.channel_dict[ch]
                         if flag[0] == 'C' and int(flag[-1]) <= self.analog_channels:
                             #POSition
-                            self.device_write(str(flag) + ':OFFSet ' + str(int(val/coef)))
+                            self.device_write(str(flag) + ':OFFSet ' + str(float(val/coef)))
 
             elif len(channel) == 1:
                 ch = str(channel[0])
